Commented out Logging
Commented out logging due to update,manage causing a conflict with the logfile which would result in a failure to update the tool. While the update and hhit are running, they are two processes trying to write to the logfile and delete, paste the logfile which is not possible in Python between two processes. Recommend to change the update, manage to not run the tool while its replacing the tool's files.