Newdev (#13)
* feat: new virtual env for hhit only

Added a trimmed down version of hazpy that includes connection string file and changed username, this allows HHIT to be updated without impacting other tools like FAST and EXPORT, however it is decoupled from the hazpy package on anaconda.org but uses code copied from the github hazpy repo; removed hazpy install in virtual environment; new virtual environment just for HHIT; disabled hazpy update check as its no longer in this specific virtual environment; set virtual env to use python 3.9 from 3.7; added pdfrw library.

* fix: condaChannel adjustment

* fix: removed whitespace and updated openssl version from i to k

* refactor: added back the hazpy dependency and hazpy checkfor updates

I found that disabling the "manage.checkForHazPyUpdates()" in update.py does not work since in manage.py under "def startApp(self, app_path, update_path):" as the 'res' variable is 0 whether the virtual environment from the config.json exists or not, so it skips the "if res != 0" that would have installed hazpy.

And the manage.py appears to rely on the "def checkForHazPyUpdates(self):" to install hazpy whether the virtual env existed or not.

I'm trying it out on the fema machine with the hazpy dependency in the environment.yaml and the update.py retaining the checkforhazpy updates. I suppose FHIT could still have hazpy in its environment and it could be updated as hazpy is updated.

* fix: added print statement to replace 'pass' in try, except

* refactor: reverted change to create separate virtual env for hhit only

HHIT will use hazus_env

* feat: added check for auth zipfile download and safe_unzip function

The check for the auth zipfile is unique to HHIT and needs to line up with the config.json zipfile url. The safe_unzip function will ensure that the files being unzipped are not being unzipped to other directories but only into the current working directory.

* fix: moved safe_unzip out of updateTool and added self variable

* refactor: reverted/removed safe_unzip function and use

* feat: Added Hazus icon; Renamed Hurrvac to Hurricane and added Hurrevac subtitle

* feat: add safe_unzip function and replaced extractall with it

* refactor: rename starting script file

hazus-hurrevac-import-tool.py was the previous name

* chore: remove commented code

* fix: changed name (#12)

Changed name from Hazus Hurrevac Import Tool to Hurricane Hazard Import Tool