Newdev (#15)
* feat: new virtual env for hhit only

Added a trimmed down version of hazpy that includes connection string file and changed username, this allows HHIT to be updated without impacting other tools like FAST and EXPORT, however it is decoupled from the hazpy package on anaconda.org but uses code copied from the github hazpy repo; removed hazpy install in virtual environment; new virtual environment just for HHIT; disabled hazpy update check as its no longer in this specific virtual environment; set virtual env to use python 3.9 from 3.7; added pdfrw library.

* fix: condaChannel adjustment

* fix: removed whitespace and updated openssl version from i to k

* refactor: added back the hazpy dependency and hazpy checkfor updates

I found that disabling the "manage.checkForHazPyUpdates()" in update.py does not work since in manage.py under "def startApp(self, app_path, update_path):" as the 'res' variable is 0 whether the virtual environment from the config.json exists or not, so it skips the "if res != 0" that would have installed hazpy.

And the manage.py appears to rely on the "def checkForHazPyUpdates(self):" to install hazpy whether the virtual env existed or not.

I'm trying it out on the fema machine with the hazpy dependency in the environment.yaml and the update.py retaining the checkforhazpy updates. I suppose FHIT could still have hazpy in its environment and it could be updated as hazpy is updated.

* fix: added print statement to replace 'pass' in try, except

* refactor: reverted change to create separate virtual env for hhit only

HHIT will use hazus_env

* feat: added check for auth zipfile download and safe_unzip function

The check for the auth zipfile is unique to HHIT and needs to line up with the config.json zipfile url. The safe_unzip function will ensure that the files being unzipped are not being unzipped to other directories but only into the current working directory.

* fix: moved safe_unzip out of updateTool and added self variable

* refactor: reverted/removed safe_unzip function and use

* feat: Added Hazus icon; Renamed Hurrvac to Hurricane and added Hurrevac subtitle

* feat: add safe_unzip function and replaced extractall with it

* refactor: rename starting script file

hazus-hurrevac-import-tool.py was the previous name

* chore: remove commented code

* fix: changed name (#12)

Changed name from Hazus Hurrevac Import Tool to Hurricane Hazard Import Tool

* Squashed commit of the following:

commit 5444bc1ce563b18be2656a95c2510365204efe1d
Author: clindeman <clindeman@niyamit.com>
Date:   Mon Jun 21 15:35:23 2021 -1000

    fix: updated requirements

commit 59b40c34ec0fb0fe0cb7fab07ae9df7ddc5ee5d1
Author: clindeman <clindeman@niyamit.com>
Date:   Wed Jun 16 12:14:44 2021 -1000

    fix: changed formatting

commit 66fa9c38e2a2280b3751a679addaee03981d47f3
Author: clindeman <clindeman@niyamit.com>
Date:   Wed Jun 16 11:34:52 2021 -1000

    fix: cleanup of content and change to markdown

commit aaceacf126b08ff10cdaa9d7771bcdc93aac3ce8
Author: clindeman <clindeman@niyamit.com>
Date:   Wed Jun 16 10:33:04 2021 -1000

    fix: added missing step

commit fa906d2bd166c8607c48b76e4c4f4bee500fc8ee
Author: clindeman <clindeman@niyamit.com>
Date:   Wed Jun 16 10:19:02 2021 -1000

    fix: added troubleshooting section

commit ecd88adeb4cae623279d56d58d882344111a7eee
Author: clindeman <clindeman@niyamit.com>
Date:   Wed Jun 16 09:59:22 2021 -1000

    fix: added instructions on what to do if updating finds many conflicts. Renamed HITT to HHIT

commit a99bb9ddbf1e006bbd2f5be538e7114d53e733b0
Author: clindeman <clindeman@niyamit.com>
Date:   Tue Jun 15 12:10:16 2021 -1000

    fix: changed name

    Changed name from Hazus Hurrevac Import Tool to Hurricane Hazard Import Tool

* Squashed commit of the following:

commit 5f7a75ef6c320a8bd0bfaedfe7ec723754749c3c
Author: clindeman <clindeman@niyamit.com>
Date:   Tue Jun 22 08:50:37 2021 -1000

    refactor: Check whoami to set proxy instead of waiting for timeout

    Also removed createProxyEnv method as it looks like it was not used.

commit 6ee076ca433f5452a642a0fca606a45f9ecddc8d
Author: Lindeman <0107902465@FEMA.DHS.GOV>
Date:   Mon Jun 21 19:16:01 2021 -1000

    fix: refactor code to work

    handleprody() now occurs before trying to access the url.

commit 1493b368fae7859db14e11643555ad8986733ae2
Author: clindeman <clindeman@niyamit.com>
Date:   Mon Jun 21 16:45:35 2021 -1000

    feat-Added Manage class to handle proxy needs